refactor(postprocess): rename scripts to concise filenames
